requirements_video.txt
torch>=2.0.0
diffusers>=0.31.0
transformers>=4.40.0
accelerate>=0.30.0
opencv-python>=4.8.0
Pillow>=10.0.0
sentencepiece>=0.2.0
protobuf>=3.20.0
numpy>=1.24.0
